Move init and update to ReduceUtils.h
