+++ aten/src/ATen/native/cpu/ReduceUtils.h
@@ -6,6 +6,7 @@
 #include <ATen/cpu/vec/functional.h>
 #include <ATen/native/ReductionType.h>
 #include <c10/util/irange.h>
+#include <ATen/OpMathType.h>
 
 namespace at::native {
 inline namespace CPU_CAPABILITY {
@@ -93,6 +94,15 @@ inline void init(scalar_t* out, int64_t size, bool include_self = false) {
   }
 }
 
+template <typename scalar_t, ReductionType reduce>
+inline void _init(scalar_t* self_ptr, at::opmath_type<scalar_t>* buffer_ptr, int64_t size, bool include_self) {
+  if (!include_self) {
+    init<at::opmath_type<scalar_t>, reduce>(buffer_ptr, size, include_self);
+  } else {
+    vec::convert(self_ptr, buffer_ptr, size);
+  }
+}
+
 template <typename scalar_t>
 inline scalar_t _max(const scalar_t& x, const scalar_t& y) {
   return at::_isnan(y) ? y : std::max(x, y);
@@ -140,6 +150,37 @@ inline void update(scalar_t* out, scalar_t* data, int64_t K) {
       out,
       data,
       K);
+}
+
+template <typename scalar_t, ReductionType reduce,
+          typename std::enable_if_t<!is_reduced_floating_point_v<scalar_t>, int> = 0>
+inline void _update(scalar_t* src_data, at::opmath_type<scalar_t>* buffer_ptr, int64_t col, int64_t K) {
+  update<scalar_t, reduce>(buffer_ptr, src_data + col * K, K);
+}
+
+template <typename scalar_t, ReductionType reduce,
+          typename std::enable_if_t<is_reduced_floating_point_v<scalar_t>, int> = 0>
+inline void _update(scalar_t* src_data, at::opmath_type<scalar_t>* buffer_ptr, int64_t col, int64_t K) {
+  using opmath_t = at::opmath_type<scalar_t>;
+  using Vec = vec::Vectorized<scalar_t>;
+  using fVec = vec::Vectorized<opmath_t>;
+  int64_t k = 0;
+  constexpr int64_t kVecSize = Vec::size();
+  constexpr int64_t kfVecSize = fVec::size();
+  for (k = 0; k < K - (K % kVecSize); k += kVecSize) {
+    Vec src_vec = Vec::loadu(src_data + col * K + k);
+    fVec src_vec0, src_vec1;
+    std::tie(src_vec0, src_vec1) = convert_to_float<scalar_t>(src_vec);
+    fVec buf_vec0, buf_vec1;
+    buf_vec0 = update<fVec, reduce>(fVec::loadu(buffer_ptr + k), src_vec0);
+    buf_vec1 = update<fVec, reduce>(fVec::loadu(buffer_ptr + k + kfVecSize), src_vec1);
+    buf_vec0.store(buffer_ptr + k);
+    buf_vec1.store(buffer_ptr + k + fVec::size());
+  }
+  for (; k < K; k++) {
+    opmath_t src_val = opmath_t(src_data[col * K + k]);
+    buffer_ptr[k] = update<opmath_t, reduce>(buffer_ptr[k], src_val);
+  }
 }
 
 template <typename scalar_t, ReductionType reduce>
